--- src/main/java/Exceptions/DukeException.java
@@ -1,7 +1,0 @@
-package Exceptions;
-
-/**
- * abstract class for custom exceptions
- */
-public abstract class DukeException extends Exception {
-}
