+++ src/main/java/commands/ToDoCommand.java
@@ -1,0 +1,34 @@
+package commands;
+
+import exceptions.MissingDescriptionException;
+import storage.Storage;
+import tasklist.TaskList;
+import ui.Ui;
+
+/**
+ * Command Object that handles creating a new ToDo
+ */
+public class ToDoCommand extends Command {
+    /**
+     * Constructs a Todo Command object
+     * @param rawUserInput
+     */
+    public ToDoCommand(String rawUserInput) {
+        super(rawUserInput);
+    }
+
+    /**
+     * Creates a new ToDo object
+     * Updates Task list and writes to Storage
+     * @param taskList object handling task operations
+     * @param ui object handling default user interface messages
+     * @param storage object handling storage operations
+     * @throws MissingDescriptionException throws when user fails to provide a description
+     */
+    @Override
+    public void execute(TaskList taskList, Ui ui, Storage storage) throws MissingDescriptionException {
+        String[] splitCommands = removeCommandWord(super.rawUserInput);
+        taskList.addNewToDo(splitCommands[1]);
+        storage.saveToHardDisk(taskList);
+    }
+}
