+++ src/main/java/exceptions/DukeException.java
@@ -1,0 +1,7 @@
+package exceptions;
+
+/**
+ * abstract class for custom exceptions
+ */
+public abstract class DukeException extends Exception {
+}
